pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gloggur"
version = "0.1.0"
description = "Symbol-level, incremental codebase indexer for semantic search and precedent retrieval."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{name = "Gloggur Contributors"}]
keywords = ["code-search", "indexer", "tree-sitter", "embeddings", "cli"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent"
]

dependencies = [
  "click>=8.1",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "watchfiles>=0.22.0",
  "tree-sitter>=0.25",
  "tree-sitter-language-pack>=0.13.0",
  "faiss-cpu>=1.8",
  "numpy>=1.26",
  "sentence-transformers>=2.7",
]

[project.optional-dependencies]
local = []
openai = ["openai>=1.3", "tenacity>=8.3"]
gemini = ["google-genai>=0.5"]
all = [
  "sentence-transformers>=2.7",
  "openai>=1.3",
  "tenacity>=8.3",
  "google-genai>=0.5",
]
dev = [
  "build>=1.2",
  "pytest>=8.2.0",
  "pytest-xdist>=3.6.1",
  "pytest-cov>=5.0.0",
  "black>=24.4.0",
  "mypy>=1.10.0",
  "ruff>=0.4.0",
]

[project.scripts]
gloggur = "gloggur.cli.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["gloggur*"]

[tool.ruff]
line-length = 100
extend-exclude = [".claude", ".gloggur-cache", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.black]
line-length = 100
extend-exclude = "/(\\.claude|\\.gloggur-cache|build|dist)/"

[tool.mypy]
python_version = "3.10"
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-n auto --dist=loadscope --cov=src/gloggur --cov-report=term-missing --cov-report=xml"
